requirements.txt
# Campus-to-Career — Backend
python-dotenv>=1.0
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg2-binary>=2.9
gunicorn>=21.0

# ML
scikit-learn>=1.3
joblib>=1.3
# Optional: NLTK, spaCy, SHAP for NLP and explainability
# nltk>=3.8
# spacy>=3.7
# shap>=0.43
